refactor: create file for remove const keysAll
